6 - Final Project/backend/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
django-cors-headers
djangorestframework-simplejwt
django-phonenumber-field[phonenumberslite]